requirements.txt
# Core
torch>=2.0.0
numpy>=1.26.0
Pillow==10.1.0

# Diffusion models
diffusers
transformers
accelerate
safetensors
huggingface_hub

# 3D processing
omegaconf>=2.3.0
einops==0.7.0
PyMCubes
trimesh>=4.0.5
imageio[ffmpeg]
xatlas==0.0.9
moderngl==5.10.0

# Background removal
rembg
onnxruntime-gpu

# API Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart

# Translation (Japanese -> English for SDXL)
deep-translator>=1.11.0

# Frontend - Gradio
gradio>=4.0.0

# Frontend - Streamlit
streamlit>=1.28.0

# Frontend - Reflex
reflex>=0.4.0

# Testing
pytest>=7.0.0
httpx>=0.24.0
